Fix rate limit issues with floats

+++ discord/http.py
@@ -941,7 +941,7 @@ class HTTPClient:
                         if isinstance(data, str):
                             # Cloudflare ban
                             is_global = False
-                            retry_after = int(response.headers.get('Retry-After', '0'))
+                            retry_after = float(response.headers.get('Retry-After', '0'))
                             if not retry_after:
                                 # Unhandleable
                                 result = _CLOUDFLARE_REGEX.search(data)
@@ -949,7 +949,7 @@ class HTTPClient:
                                 raise HTTPException(response, f'Cloudflare ban (code: {code})')
                         else:
                             is_global: bool = data.get('global', False)
-                            retry_after: float = data.get('retry_after', int(response.headers.get('Retry-After', 0)))
+                            retry_after: float = data.get('retry_after', float(response.headers.get('Retry-After', 0)))
 
                         # Cloudflare rate limit
                         is_cloudflare = not response.headers.get('Via')
@@ -969,7 +969,7 @@ class HTTPClient:
 
                         if 'Retry-After' in response.headers:
                             # Sometimes Cloudflare rate limits will have their retry_after field in milliseconds
-                            if int(response.headers['Retry-After']) == int(retry_after / 1000):  # type: ignore
+                            if float(response.headers['Retry-After']) == retry_after / 1000:  # type: ignore
                                 retry_after /= 1000.0
 
                         if self.max_ratelimit_timeout and retry_after > self.max_ratelimit_timeout:
